mqtt_client/requirements.txt
paho-mqtt==1.6.1
dash==2.14.1
pandas==2.1.1
python-dotenv==1.0.0
